Add Gitlab::QA::Docker::Gitlab#environment accessor
In order to enable a configuration of the gitlab omnibus by environment
variables such as `GITLAB_OMNIBUS_CONFIG` we add an #environment
accessor and make use of it in the #start method.

Adds rspec testing of Docker::Gitlab#start, including that the new
parameter is reflected in the generated command.