Add a Test::Sanity::Version suite
Pull an image, looks at the version-manifest to find what the sha is for
gitlab-rails, and checks it is a commit that was published to
gitlab-org/gitlab-{ce,ee} in a sampling of previously published commits.

skopeo (or other api client) and Labels would reduce network traffic for
the pull, but it's not uncommon we want to use what we pulled anyhow.

* Adds `Engine#read_file(image, tag, path)` to get a file contents from the image
* Adds `Component::Gitlab#pull` to allow for us to pull and check more fluently
* Adds `Component::Gitlab#sha_version` to get the version of `gitlab-rails`
* Adds `Release#project_name` to allow us to find a project for a release

Return output from Shellout.execute!

Here we add a defined return value to Shellout.execute!, it's the STDOUT
from the executed command.  This allows us to be more terse in
Component::Gitlab#sha_version as the block supplied to Engine#read_file
doesn't need to collect the lines, it's now just there to keep the
command output quiet.

Supress rubocop warnings

Methods and Classes get long.

Don't supress printing of `version-manifest.json`

In favour of terser code, allow the `version-manifest` to go to STDOUT.