Change the contract of Shellout.execute!
When a block is supplied, it's the callers responsibility to handle
everything to do with the line, be that printing it or just collecting
it up for later user.  If the block was not supplied we continue to
print the line as it has some debug value.

We update the use in Component::Gitlab#reconfigure to continue printing
the line as it has value.

Additionally we change the return value to be the collected output lines
from the command, so helpers like read_file don't need to collect lines
up.