Add basic LDAP integration tests
This commit adds an LDAP component that sets up the docker-openldap
(https://github.com/osixia/docker-openldap) image with the proper
configuration for GitLab users to login.

By default, the docker-openldap image configures the Docker image with a
default admin user in the example.org domain. This user does not have a uid
attribute that GitLab needs to authenticate, so we seed the LDAP server with a
"tanuki" user via a LDIF file in the fixtures/ldap directory.