Skip tests during previous_release stage of Update scenario
Assuming the previous release has already been tested, testing it again
is inefficient and troublesome when a failing test has been fixed in the
current release but keeps failing when run in the first stage of
this Update scenario.

This change launches an instance with the previous release so that the
temp volumes are created and can be used by the current release, but it
doesn't run the tests on the previous release.