Add QA framework OptionParser options
The options need to be specified here otherwise gitlab-qa treats
them as invalid options.

We pass all the options through to the QA framework. This avoids the
problem of needing to make sure there's only one `--` separator; we
pass through all the options, including the separator.

This also makes exe/gitlab-qa redundant with bin/qa so we don't
have to maintain duplicate scripts.